+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -1842,12 +1842,12 @@ export class PortfolioService {
     userId: string;
     withExcludedAccounts?: boolean;
   }) {
-    const ordersOfTypeItem = await this.orderService.getOrders({
+    const ordersOfTypeItemOrLiability = await this.orderService.getOrders({
       filters,
       userCurrency,
       userId,
       withExcludedAccounts,
-      types: ['ITEM']
+      types: ['ITEM', 'LIABILITY']
     });
 
     const accounts: PortfolioDetails['accounts'] = {};
@@ -1887,13 +1887,14 @@ export class PortfolioService {
         return accountId === account.id;
       });
 
-      const ordersOfTypeItemByAccount = ordersOfTypeItem.filter(
-        ({ accountId }) => {
+      const ordersOfTypeItemOrLiabilityByAccount =
+        ordersOfTypeItemOrLiability.filter(({ accountId }) => {
           return accountId === account.id;
-        }
-      );
+        });
 
-      ordersByAccount = ordersByAccount.concat(ordersOfTypeItemByAccount);
+      ordersByAccount = ordersByAccount.concat(
+        ordersOfTypeItemOrLiabilityByAccount
+      );
 
       accounts[account.id] = {
         balance: account.balance,
@@ -1933,7 +1934,7 @@ export class PortfolioService {
             order.unitPrice ??
             0);
 
-        if (order.type === 'SELL') {
+        if (order.type === 'LIABILITY' || order.type === 'SELL') {
           currentValueOfSymbolInBaseCurrency *= -1;
         }